requirements.txt
gTTS
ollama-python
requests
beautifulsoup4
google-api-python-client
openai-whisper
SpeechRecognition
pygame
pyaudio
tkinter
langchain
langchain_community
